+++ Marlin/src/lcd/extui/dgus_lcd.cpp
@@ -21,19 +21,19 @@
  */
 
 /**
- * extui_dgus_lcd.cpp
+ * dgus_lcd.cpp
  *
  * DGUS implementation for Marlin by coldtobi, Feb-May 2019
  */
 
-#include "../inc/MarlinConfigPre.h"
+#include "../../inc/MarlinConfigPre.h"
 
 #if HAS_DGUS_LCD
 
-#include "extui/ui_api.h"
-#include "extui/lib/dgus/DGUSDisplay.h"
-#include "extui/lib/dgus/DGUSDisplayDef.h"
-#include "extui/lib/dgus/DGUSScreenHandler.h"
+#include "ui_api.h"
+#include "lib/dgus/DGUSDisplay.h"
+#include "lib/dgus/DGUSDisplayDef.h"
+#include "lib/dgus/DGUSScreenHandler.h"
 
 extern const char NUL_STR[];
