+++ platformio.ini
@@ -49,12 +49,12 @@ default_src_filter = +<src/*> -<src/config> -<src/HAL> +<src/HAL/shared>
   -<src/lcd/menu/menu_touch_screen.cpp>
   -<src/lcd/menu/menu_ubl.cpp>
   -<src/lcd/extui/lib/mks_ui>
-  -<src/lcd/extui/lib/dgus> -<src/lcd/extui_dgus_lcd.cpp>
-  -<src/lcd/extui_example.cpp>
-  -<src/lcd/extui_malyan_lcd.cpp>
+  -<src/lcd/extui/lib/dgus> -<src/lcd/extui/dgus_lcd.cpp>
+  -<src/lcd/extui/example.cpp>
+  -<src/lcd/extui/malyan_lcd.cpp>
   -<src/lcd/extui/lib/ftdi_eve_touch_ui>
   -<src/lcd/extui/lib/anycubic_i3mega>
-  -<src/lcd/extui_anycubic_tft.cpp>
+  -<src/lcd/extui/anycubic_tft.cpp>
   -<src/lcd/lcdprint.cpp>
   -<src/sd/usb_flashdrive>
   -<src/feature/backlash.cpp>
@@ -241,11 +241,11 @@ HAS_MENU_TMC            = src_filter=+<src/lcd/menu/menu_tmc.cpp>
 HAS_MENU_TOUCH_SCREEN   = src_filter=+<src/lcd/menu/menu_touch_screen.cpp>
 HAS_MENU_UBL            = src_filter=+<src/lcd/menu/menu_ubl.cpp>
 ANYCUBIC_LCD_I3MEGA     = src_filter=+<src/lcd/extui/lib/anycubic_i3mega>
-HAS_DGUS_LCD            = src_filter=+<src/lcd/extui/lib/dgus> +<src/lcd/extui_dgus_lcd.cpp>
+HAS_DGUS_LCD            = src_filter=+<src/lcd/extui/lib/dgus> +<src/lcd/extui/dgus_lcd.cpp>
 TOUCH_UI_FTDI_EVE       = src_filter=+<src/lcd/extui/lib/ftdi_eve_touch_ui>
-HAS_ANYCUBIC_TFT_EXTUI  = src_filter=+<src/lcd/extui_anycubic_tft.cpp>
-EXTUI_EXAMPLE           = src_filter=+<src/lcd/extui_example.cpp>
-MALYAN_LCD              = src_filter=+<src/lcd/extui_malyan_lcd.cpp>
+HAS_ANYCUBIC_TFT_EXTUI  = src_filter=+<src/lcd/extui/anycubic_tft.cpp>
+EXTUI_EXAMPLE           = src_filter=+<src/lcd/extui/example.cpp>
+MALYAN_LCD              = src_filter=+<src/lcd/extui/malyan_lcd.cpp>
 HAS_SPI_LCD             = src_filter=+<src/lcd/lcdprint.cpp>
 USB_FLASH_DRIVE_SUPPORT = src_filter=+<src/sd/usb_flashdrive>
 AUTO_BED_LEVELING_BILINEAR = src_filter=+<src/feature/bedlevel/abl>
